Extract many things
These were all tool-assisted, so should be relatively safe.

TODO: Check package consumers for previously exported types, etc.
