+++ src/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_QUERY_LIMIT = 50;
+export const MAX_BATCH_OPERATIONS = 25;
