+++ test/batchWrite.test.ts
@@ -1,6 +1,6 @@
 import partition from 'lodash.partition';
 import { v4 as uuid } from 'uuid';
-import { BatchWriteOperation } from '../src';
+import { BatchWriteOperation } from '../src/types';
 import TestContext, {
   DataModel,
   documentClient,
